remove score_sde_pytorch namespace
no reason anymore to have it in this namespace now that parallel deterministic namespace has been removed
