+++ .pre-commit-config.yaml
@@ -13,9 +13,9 @@ repos:
     hooks:
       - id: black
         language_version: python3.9
-        exclude: ^src/ml_downscaling_emulator/score_sde_pytorch/
+        exclude: ^src/ml_downscaling_emulator/(run_lib.py|sde_lib.py|likelihood.py|sampling.py|losses.py|models|op|configs)
   - repo: https://github.com/pycqa/flake8
     rev: '6.0.0'  # pick a git hash / tag to point to
     hooks:
       - id: flake8
-        exclude: ^src/ml_downscaling_emulator/score_sde_pytorch/
+        exclude: ^src/ml_downscaling_emulator/(run_lib.py|sde_lib.py|likelihood.py|sampling.py|losses.py|models|op|configs)
