+++ src/ml_downscaling_emulator/configs/deterministic/ukcp_local_pr_12em_plain_unet.py
@@ -15,9 +15,13 @@
 # limitations under the License.
 
 # Lint as: python3
-"""Training NCSN++ on precip data in a deterministic fashion."""
+"""Debug config for training a purely deterministic model.
 
-from ml_downscaling_emulator.score_sde_pytorch.configs.deterministic.default_configs import get_default_configs
+This is opposed to using a model ready for score-based denoising
+but training it in a deterministic fashion.
+"""
+
+from ml_downscaling_emulator.configs.deterministic.default_configs import get_default_configs
 
 def get_config():
   config = get_default_configs()
@@ -25,38 +29,29 @@ def get_config():
   # training
   training = config.training
   training.n_epochs = 100
+  training.snapshot_freq = 20
+  training.batch_size = 64
 
   # data
   data = config.data
   data.dataset_name = 'bham64_ccpm-4x_12em_psl-sphum4th-temp4th-vort4th_pr'
+  data.input_transform_key = "stan"
+  data.target_transform_key = "sqrturrecen"
+  data.input_transform_dataset = None
+  data.time_inputs = False
 
   # model
   model = config.model
-  model.name = 'cncsnpp'
-  model.loc_spec_channels = 0
-  model.dropout = 0.1
-  model.embedding_type = 'fourier'
-  model.scale_by_sigma = False
-  model.ema_rate = 0.9999
-  model.normalization = 'GroupNorm'
-  model.nonlinearity = 'swish'
-  model.nf = 128
-  model.ch_mult = (1, 2, 2, 2)
-  model.num_res_blocks = 4
-  model.attn_resolutions = (16,)
-  model.resamp_with_conv = True
-  model.conditional = True
-  model.fir = True
-  model.fir_kernel = [1, 3, 3, 1]
-  model.skip_rescale = True
-  model.resblock_type = 'biggan'
-  model.progressive = 'none'
-  model.progressive_input = 'residual'
-  model.progressive_combine = 'sum'
-  model.attention_type = 'ddpm'
-  model.embedding_type = 'positional'
-  model.init_scale = 0.
-  model.fourier_scale = 16
-  model.conv_size = 3
+  model.name = 'det_cunet'
+  model.ema_rate = 1 # basically disables EMA
 
+  # optimizer
+  optim = config.optim
+  optim.optimizer = "Adam"
+  optim.lr = 2e-4
+  optim.beta1 = 0.9
+  optim.eps = 1e-8
+  optim.weight_decay = 0
+  optim.warmup = -1 # 5000
+  optim.grad_clip = -1. # 1.
   return config
