+++ src/ml_downscaling_emulator/configs/subvpsde/ukcp_local_mv_debug.py
@@ -16,7 +16,7 @@
 # Lint as: python3
 """Training conditional U-Net on precip data with sub-VP SDE.
 DEBUGGING ONLY"""
-from ml_downscaling_emulator.score_sde_pytorch.configs.default_ukcp_local_pr_1em_configs import get_default_configs
+from ml_downscaling_emulator.configs.default_ukcp_local_pr_1em_configs import get_default_configs
 
 
 def get_config():
