+++ src/ml_downscaling_emulator/configs/subvpsde/ukcp_local_pr_12em_cncsnpp_continuous.py
@@ -15,7 +15,7 @@
 
 # Lint as: python3
 """Training NCSN++ on precip data with sub-VP SDE."""
-from ml_downscaling_emulator.score_sde_pytorch.configs.default_ukcp_local_pr_1em_configs import get_default_configs
+from ml_downscaling_emulator.configs.default_ukcp_local_pr_12em_configs import get_default_configs
 
 
 def get_config():
@@ -64,7 +64,5 @@ def get_config():
 
   # data
   data = config.data
-  data.input_transform_key = "stan"
-  data.target_transform_key = "sqrturrecen"
 
   return config
