+++ src/ml_downscaling_emulator/configs/subvpsde/ukcp_local_pr_1em_cncsnpp_continuous.py
@@ -15,7 +15,7 @@
 
 # Lint as: python3
 """Training NCSN++ on precip data with sub-VP SDE."""
-from ml_downscaling_emulator.score_sde_pytorch.configs.default_ukcp_local_pr_12em_configs import get_default_configs
+from ml_downscaling_emulator.configs.default_ukcp_local_pr_1em_configs import get_default_configs
 
 
 def get_config():
@@ -35,7 +35,6 @@ def get_config():
   # data
   data = config.data
   data.centered = True
-  data.dataset_name = 'bham64_ccpm-4x_12em_psl-sphum4th-temp4th-vort4th_tmean150cm'
 
   # model
   model = config.model
@@ -65,6 +64,7 @@ def get_config():
 
   # data
   data = config.data
-  data.target_transform_key = 'mm;recen'
+  data.input_transform_key = "stan"
+  data.target_transform_key = "sqrturrecen"
 
   return config
