+++ src/ml_downscaling_emulator/configs/subvpsde/ukcp_local_pr_1em_cncsnpp_continuous_ld.py
@@ -15,7 +15,7 @@
 
 # Lint as: python3
 """Training NCSN++ on precip data with sub-VP SDE."""
-from ml_downscaling_emulator.score_sde_pytorch.configs.default_ukcp_local_pr_12em_configs import get_default_configs
+from ml_downscaling_emulator.configs.default_ukcp_local_pr_1em_configs import get_default_configs
 
 
 def get_config():
@@ -25,6 +25,7 @@ def get_config():
   training.sde = 'subvpsde'
   training.continuous = True
   training.reduce_mean = True
+  training.n_epochs = 300
 
   # sampling
   sampling = config.sampling
@@ -35,8 +36,7 @@ def get_config():
   # data
   data = config.data
   data.centered = True
-  data.dataset_name = 'bham64_ccpm-4x_12em_mv'
-  data.target_transform_overrides.target_tmean150cm = "mm;recen"
+  data.dataset_name = 'bham64_ccpm-4x_1em_psl-sphum4th-temp4th-vort4th_pr-historic'
 
   # model
   model = config.model
@@ -66,5 +66,7 @@ def get_config():
 
   # data
   data = config.data
+  data.input_transform_key = "stan"
+  data.target_transform_key = "sqrturrecen"
 
   return config
