+++ src/ml_downscaling_emulator/configs/vesde/ukcp_local_pr_cunet_continuous.py
@@ -15,7 +15,7 @@
 
 # Lint as: python3
 """Training UNet on XArray with VE SDE."""
-from ml_downscaling_emulator.score_sde_pytorch.configs.default_ukcp_local_pr_1em_configs import get_default_configs
+from ml_downscaling_emulator.configs.default_ukcp_local_pr_1em_configs import get_default_configs
 
 
 def get_config():
