+++ tests/smoke-tests/test-det-debug-cunet
@@ -5,7 +5,7 @@ set -euo pipefail
 config_name="ukcp_local_pr_debug"
 
 workdir="output/test/deterministic/${config_name}/test-run"
-config_path="src/ml_downscaling_emulator/score_sde_pytorch/configs/deterministic/${config_name}.py"
+config_path="src/ml_downscaling_emulator/configs/deterministic/${config_name}.py"
 
 loc_spec_channels=0
