+++ tox.ini
@@ -1,5 +1,5 @@
 [flake8]
 max-line-length = 88
 extend-ignore = E203,E501,F541
-exclude = .git,__pycache__,docs/source/conf.py,old,build,dist,src/ml_downscaling_emulator/unet,src/ml_downscaling_emulator/score_sde_pytorch
+exclude = .git,__pycache__,docs/source/conf.py,old,build,dist,src/ml_downscaling_emulator/unet,src/ml_downscaling_emulator/run_lib.py,src/ml_downscaling_emulator/sde_lib.py,src/ml_downscaling_emulator/likelihood.py,src/ml_downscaling_emulator/sampling.py,src/ml_downscaling_emulator/losses.py,src/ml_downscaling_emulator/models, src/ml_downscaling_emulator/op,src/ml_downscaling_emulator/configs
 max-complexity = 20
